pyproject.toml
[project]
name = "observatory-booking"
version = "0.1.0"
description = "A Flask-based, self-hostable observatory booking web app with role-based access control, weather-aware scheduling, and encrypted user data."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "apscheduler>=3.11.2",
    "bcrypt>=5.0.0",
    "cryptography>=46.0.3",
    "python-dotenv>=1.0.0",
    "flask>=3.1.2",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.45",
    "waitress>=3.0.2",
    "werkzeug>=3.1.5",
]

[project.optional-dependencies]
dev = [
    "black>=25.12.0",
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "pylint>=4.0.4",
    "types-requests>=2.32.4.20260107",
    "types-waitress>=3.0.1.20250801",
]

docs = [
    "sphinx>=7.4.7",
    "pydata-sphinx-theme>=0.16.1",
    "sphinxcontrib-httpdomain>=1.8.1",
    "sphinx-copybutton>=0.5.2",
    "sphinxcontrib-openapi>=0.8.4",
]

[tool.pylint.MASTER]
init-hook = "import sys, os, glob; sp = glob.glob(os.path.join('.venv', 'Lib', 'site-packages')) + glob.glob(os.path.join('.venv', 'lib', 'python*', 'site-packages')); [sys.path.insert(0, p) for p in sp if p not in sys.path]"
